backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch>=2.6.0
pandas==2.1.3
numpy==1.25.2
xarray==2023.11.0
netcdf4==1.6.5
pymongo==4.6.0
motor==3.3.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic>=2.5.0
scikit-learn==1.3.2
plotly==5.17.0
kaleido==0.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6